modified functions so all actions are within analyzeuser view
